+++ tools/release.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2025 Luca Baldini (luca.baldini@pi.infn.it)
+# Copyright 2023--2025 the aptapy team
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by