+++ docs/release_notes.rst
@@ -3,6 +3,14 @@
 Release notes
 =============
 
+* Changed parametrization of the PowerLaw model to include an arbitrary pivot that
+  can be set at creation time.
+* Minor warning fixes in various places (docs and unit tests).
+* Pull requests merged and issues closed:
+
+  - https://github.com/lucabaldini/aptapy/pull/80
+  - https://github.com/lucabaldini/aptapy/issues/79
+
 
 Version 0.17.2 (2025-12-11)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~