+++ docs/release_notes.rst
@@ -5,9 +5,13 @@ Release notes
 
 * Changed parametrization of the PowerLaw model to include an arbitrary pivot that
   can be set at creation time.
+* Probit fitting model added.
+* Minor warning fixes in various places (docs and unit tests).
 * Pull requests merged and issues closed:
 
   - https://github.com/lucabaldini/aptapy/pull/80
+  - https://github.com/lucabaldini/aptapy/pull/82
+  - https://github.com/lucabaldini/aptapy/issues/81
   - https://github.com/lucabaldini/aptapy/issues/79