+++ docs/release_notes.rst
@@ -7,6 +7,7 @@ Release notes
   can be set at creation time.
 * Probit fitting model added.
 * Added the Histogram3d class for three-dimensional histograms.
+* New slice1d() method in the histogram base class.
 * Minor warning fixes in various places (docs and unit tests).
 * Pull requests merged and issues closed: